Backend/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
bcrypt==4.2.0
boto3==1.35.33
botocore==1.35.33
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
decorator==5.1.1
executing==2.1.0
fastapi==0.112.3
greenlet==3.0.3
h11==0.14.0
idna==3.8
ipython==8.27.0
jedi==0.19.1
jmespath==1.0.1
jwcrypto==1.5.6
jwt==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
parso==0.8.4
passlib==1.7.4
prompt_toolkit==3.0.48
psycopg-binary==3.2.1
psycopg2==2.9.9
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
python-multipart==0.0.12
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.38.4
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
wcwidth==0.2.13
